+++ visualization/chat_inference.py
@@ -101,7 +101,7 @@ class ChatInference:
 
             self.model = AutoModelForCausalLM.from_pretrained(
                 model_id,
-                torch_dtype=torch.float16,
+                torch_dtype=torch.bfloat16,
                 device_map=self.device,
                 attn_implementation='eager'
             )
